fix(suite-native): Resolve circular dependency with react-native-graph types
