+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -109,4 +109,3 @@ suite-common/wallet-utils/src/index.ts > suite-common/wallet-utils/src/accountUt
 suite-native/device/src/components/ConfirmOnTrezor/ConfirmOnTrezorWrapper.tsx > suite-native/device/src/components/ConfirmOnTrezor/ConfirmOnTrezorContent.tsx
 suite-native/device/src/index.ts > suite-native/device/src/hooks/useDetectDeviceError.tsx > suite-native/device/src/selectors.ts > suite-native/firmware/src/index.ts > suite-native/firmware/src/components/FirmwareInstallationScreenContent.tsx
 suite-native/navigation/src/index.ts > suite-native/navigation/src/hooks/useNavigateToInitialScreen.ts
-suite-native/react-native-graph/src/LineGraphProps.ts > suite-native/react-native-graph/src/CreateGraphPath.ts
